deletes user without deleting default pic file

+++ Backend/services/userService.js
@@ -150,9 +150,11 @@ function deleteUser(userId) {
         throw err;
       });
       // Delete image in storage
-      fs.unlink(user.image, (err) => {
-        // in case of error, skip and continue
-      });
+      if(!(user.image.includes("default_profile"))) {
+        fs.unlink(user.image, (err) => {
+          // in case of error, skip and continue
+        });
+      }
       return resolve({
         data: user,
         message: "User wurde erfolgreich entfernt.",
